added a zoo library
